# Help LLVM vectorize comparison kernel ~50-80% faster