# Which issue does this PR close?

<!--
We generally require a GitHub issue to be filed for all bug fixes and enhancements and this helps us generate change logs for our releases. You can link an issue to this PR using the GitHub syntax. For example `Closes #123` indicates that this PR will close issue #123.
-->

Closes #.

# Rationale for this change
 
<!--
Why are you proposing this change? If this is already explained clearly in the issue then this section is not needed.
Explaining clearly why changes are proposed helps reviewers understand your changes and offer better suggestions for fixes.
-->

I was messing around in godbolt and realised these loops were not getting vectorised correctly

**Default flags**

```
eq Float32              time:   [36.531 us 36.547 us 36.566 us]                        
                        change: [-45.190% -45.147% -45.104%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 5 outliers among 100 measurements (5.00%)
  1 (1.00%) low mild
  1 (1.00%) high mild
  3 (3.00%) high severe
neq Float32             time:   [36.677 us 36.695 us 36.718 us]                         
                        change: [-48.185% -48.131% -48.071%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 12 outliers among 100 measurements (12.00%)
  2 (2.00%) low mild
  2 (2.00%) high mild
  8 (8.00%) high severe

```

**Target CPU**

```
eq Float32              time:   [19.023 us 19.040 us 19.060 us]                        
                        change: [-71.418% -71.385% -71.347%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 16 outliers among 100 measurements (16.00%)
  9 (9.00%) high mild
  7 (7.00%) high severe

neq Float32             time:   [19.172 us 19.184 us 19.197 us]                         
                        change: [-72.938% -72.919% -72.900%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 8 outliers among 100 measurements (8.00%)
  1 (1.00%) low mild
  1 (1.00%) high mild
  6 (6.00%) high severe
```

Once this is in, I will give the same treatment to the scalar kernels.

As an added bonus this likely reduces the amount of generated LLVM IR, and so may help very slightly with compile times.

# What changes are included in this PR?

<!--
There is no need to duplicate the description in the issue here but it is sometimes worth providing a summary of the individual changes in this PR.
-->

# Are there any user-facing changes?


<!--
If there are user-facing changes then we may require documentation to be updated before approving the PR.
-->

<!---
If there are any breaking changes to public APIs, please add the `breaking change` label.
-->