@tustvold 2022-09-04
Latest numbers

```
eq Float32              time:   [18.303 us 18.316 us 18.330 us]                        
                        change: [-72.558% -72.530% -72.503%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 3 outliers among 100 measurements (3.00%)
  1 (1.00%) high mild
  2 (2.00%) high severe

eq scalar Float32       time:   [7.8778 us 7.8809 us 7.8846 us]                               
                        change: [-81.777% -81.768% -81.758%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 8 outliers among 100 measurements (8.00%)
  3 (3.00%) high mild
  5 (5.00%) high severe

neq Float32             time:   [18.294 us 18.310 us 18.326 us]                         
                        change: [-74.175% -74.153% -74.132%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 5 outliers among 100 measurements (5.00%)
  5 (5.00%) high mild

neq scalar Float32      time:   [7.9744 us 7.9777 us 7.9812 us]                                
                        change: [-81.537% -81.526% -81.512%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 6 outliers among 100 measurements (6.00%)
  1 (1.00%) high mild
  5 (5.00%) high severe

lt Float32              time:   [18.281 us 18.295 us 18.310 us]                        
                        change: [-62.585% -62.538% -62.494%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 18 outliers among 100 measurements (18.00%)
  8 (8.00%) low severe
  5 (5.00%) high mild
  5 (5.00%) high severe

lt scalar Float32       time:   [7.8722 us 7.8759 us 7.8811 us]                               
                        change: [-83.746% -83.738% -83.731%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 6 outliers among 100 measurements (6.00%)
  1 (1.00%) high mild
  5 (5.00%) high severe

lt_eq Float32           time:   [18.299 us 18.309 us 18.320 us]                           
                        change: [-63.258% -63.198% -63.139%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 5 outliers among 100 measurements (5.00%)
  5 (5.00%) high severe

lt_eq scalar Float32    time:   [7.8834 us 7.8887 us 7.8943 us]                                  
                        change: [-84.493% -84.472% -84.452%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 10 outliers among 100 measurements (10.00%)
  9 (9.00%) high mild
  1 (1.00%) high severe

gt Float32              time:   [18.277 us 18.284 us 18.293 us]                        
                        change: [-62.697% -62.632% -62.586%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 6 outliers among 100 measurements (6.00%)
  1 (1.00%) low mild
  3 (3.00%) high mild
  2 (2.00%) high severe

gt scalar Float32       time:   [7.8602 us 7.8642 us 7.8689 us]                               
                        change: [-82.399% -82.389% -82.380%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 10 outliers among 100 measurements (10.00%)
  2 (2.00%) high mild
  8 (8.00%) high severe

gt_eq Float32           time:   [18.299 us 18.312 us 18.324 us]                           
                        change: [-62.749% -62.710% -62.671%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 3 outliers among 100 measurements (3.00%)
  1 (1.00%) high mild
  2 (2.00%) high severe

gt_eq scalar Float32    time:   [7.8686 us 7.8722 us 7.8763 us]                                  
                        change: [-82.149% -82.141% -82.132%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 5 outliers among 100 measurements (5.00%)
  1 (1.00%) low mild
  3 (3.00%) high mild
  1 (1.00%) high severe
```