@Dandandan 2022-09-04
Looks good! Happy about the scalar perf too. One remaining comment on the (un)safety remarks